funzioni completate (sistemare costruttori e fare test)

+++ Vinted/src/main/java/com/example/demo/service/AcquistoService.java
@@ -2,6 +2,7 @@ package com.example.demo.service;
 
 import com.example.demo.model.AcquistoModel;
 import com.example.demo.model.ArticoloModel;
+import com.example.demo.model.UtenteModel;
 import com.example.demo.repository.AcquistoRepository;
 import com.example.demo.repository.ArticoloRepository;
 import com.example.demo.repository.ArticoloWishListRepository;
@@ -26,9 +27,10 @@ public class AcquistoService {
 
     public boolean createAcquisto(AcquistoModel ac){
         for(ArticoloModel a : this.articoloRepo.getArticoli()) {
-            if (a.getId().equals(ac.getId_articolo())) {
+            if (a.getId().equals(ac.getId_articolo()) && a.isIn_vendita()) {
                 if (ac.getData_acquisto().compareTo(a.getData_pubblicazione()) >= 0 && ac.getData_acquisto().compareTo(ac.getData_spedizione()) <= 0) {
                     a.setIn_vendita(false);
+                    this.articoloRepo.updateArticolo(a);
                     this.artWishRepo.removeArticoloFromEveryWishlists(a);
                     return this.acquistoRepo.createAcquisto(ac);
                 }
@@ -45,4 +47,7 @@ public class AcquistoService {
         return this.acquistoRepo.removeAcquisto(id);
     }
     public List<AcquistoModel> getAcquisti(){ return this.acquistoRepo.getAcquisti(); }
+
+    public List<AcquistoModel> getAcquistiUtente(String id_utente){ return this.acquistoRepo.getAcquistiUtente(id_utente); }
+    public List<AcquistoModel> getAcquistiRangeTempo(int range){ return this.acquistoRepo.getAcquistiRangeTempo(range); }
 }
